Refactor purchases queries to use upgrades queryKey in api-queries
