+++ packages/api-queries/src/index.ts
@@ -93,3 +93,4 @@ export * from './site-users';
 export * from './site-wordpress-version';
 export * from './site';
 export * from './sites';
+export * from './upgrades';
